+++ lib/functions/helper-functions.php
@@ -10,14 +10,14 @@
  * @copyright    Copyright (c) 2024, Matt Ryan
  * @license      http://opensource.org/licenses/gpl-2.0.php GNU Public License
  */
-namespace capweb\license_manager;
+// namespace capweb;
 
 /**
  * Detect if Gravity Forms plugin active. 
  */
-if ( is_plugin_active( 'gravityforms/gravityforms.php' ) ) {
-    add_filter('gform_validation_1',   __NAMESPACE__ . '\validate_code');
-}
+// if ( is_plugin_active( 'gravityforms/gravityforms.php' ) ) {
+    add_filter('gform_validation_1',  'validate_code');
+// }
 function validate_code($validation_result){
 // http://www.gravityhelp.com/forums/topic/database-look-up-validation
 // validate 9 digit code