+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-02-19
+
+### Removed
+
+- Removed namespacing throughout plugin for simplicity. 
+
+## [1.0.3] - 2025-02-04
+
+### Changed
+
+- Refactor include for get_plugin_data.
+
 ## [1.0.2] - 2025-01-28
 
 ### Removed